agents/modulir/estimate_effort/requirements.txt
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.0.1
langgraph>=0.0.20
openai>=1.12.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
